Added new component to handle info links in MapSearchSingleTargetView and FavouriteCard
