+++ src/main.scss
@@ -18,6 +18,7 @@
 @import 'plotSearchAndCompetitionsPage/components/mapSearchComponent';
 @import 'plotSearchAndCompetitionsPage/components/mapSearchSingleTargetView';
 @import 'plotSearchAndCompetitionsPage/components/mapComponent';
+@import 'plotSearchAndCompetitionsPage/components/infoLinks';
 @import 'favouritesPage/favouritesPage';
 @import 'favouritesPage/components/favouriteCard';
 @import 'favouritesPage/components/favouriteCardDetails';
