+++ src/plotSearchAndCompetitionsPage/components/_mapSearchSingleTargetView.scss
@@ -33,16 +33,4 @@
     width: auto;
     margin: 0 auto;
   }
-
-  &__info-links {
-    padding: 0;
-    margin: 0;
-
-    li {
-      list-style: none;
-      padding: ($spacing-s / 2) 0;
-      font-size: $fontsize-body-s;
-      font-weight: 500;
-    }
-  }
 }
